Added expression code to the grammar -
This is stage 1 - we are able to compile but dont know if we can do
anything useful yet. The sources need a lot of cleanup and some files
like tree.[hC] have to be given names that reflect their purpose or
deleted. Also unnecessary dependencies among header files need to be
deleted.