This revision has some major changes. However ! IMPORTANT WARNING ! this is far from complete.
1) added small helper functions to idenity if an expression is a
constant expression
2) if an expression is an integral expression
3)  Added for statment syntax parsing to compiler. 

Further work that needs to be done is : 
1) a compound body statement should be able to tell if contained a
question - and if yes - set a flag
2) when constructing a for body statement - we check if the contained
for_body, which is a compound statement - because of the grammar I have
chosen, contains any questions. If so - we need to ensure that the for
test condition is an integer and a constant. Otherwise we cannot create
a fixed data map for the questionnaire. 
On the other hand if the for body does not contain a question then there
is no problem

demonstrative examples
// ---------------START OF EXAMPLEs---------------------------
Example 1: not Allowed
int32_t my_limit=10;
for(int i=0; i<my_limit; i=i+1){
       q4 "blah blah blah" .... // this is not allowed because my_limit
               // is variable
}

Example 2: Should be allowed
const int32_t my_const_limit=50;
for(int i=0; i<my_const_limit * 20; i=i+1){
       q4 "blah blah blah" .... // this is allowed because
               // my_const_limit is a constant variable and 20 is a constant
               // number so the loop counter is constant
}

Example 3: no questions in the compound body - so there is no condition
of the for loop index
int32_t my_limit=10;

        Here we do not care about the condition of the for loop as the
        compound statement body does not contain a question
                  |
                  V


int32_t my_limit=10;
for(int i=0; i<my_limit; i=i+1){
        // no questions in the cmpd body
}
// ---------------END OF EXAMPLEs---------------------------

Further notes: It does not matter at what level of nesting a compound
body contains a question - as soon as some inner stmt contains a
question the for counter has to be a constant integer - at least this is
my understanding as of now

for example:

for(int i=0; i<10; ++i){
        if (some_condn){
                q5 "ask q5" ....                
        }
}