Modified cmake to use 2.8 and the new FindBISON/FindFLEX commands.
The bison files generated when the input is *.ypp -> *.cpp *.hpp
there is no way we can intercept this in the CMAKE commands flex/bison stage
and rename the file.

I have to make one more change to my source code which is
change #include *.h to #include *.hpp for all files which include 
parser data - i.e. the flex files as they use the *.h files generated by
bison. This change has not been put in yet but will be done shortly (tomorrow?)