This revision does not generate code which will compile with the RuntimeMakefile However I checked in this code as the number of changes are massive and if I lost data due to a disk crash it would be very expensive in terms of the amount of work done.
Modified the way the push and pop variables are handled just before a question.
The parser now builds a list of active variables for each question and stores 
it in the question object. Added a new structure to hold the active variable 
information.

Some of the code generated is not correct at this point and will be fixed 
in the next few revisions