This is another milestone in Array Question navigation
I have added a new virtual function "GetQuestionsInBlock" which accumulates
all the questions in a Compound Code block and stores it into a vector which 
is a new CompoundStatement member "questionsInBlock_"

This is needed to implement the array question navigation.

Consider the following code

{
	int32_t i=0;
	for(i=0; i<5; i=i+1){
		ask q1;
	}
}
here when we reach the second iteration q1[1] (indexes start from 0 for arrays)
q1 has already been answered. Hence we should be saving q1 into the scope
under a suitably named key.

	The newly added function to the question class does this work - below for reference.
	void PrintSaveMyPreviousIterationsData(StatementCompiledCode &code);

Also consider the following case
{
	int32_t i=0,j=0;
	for(i=0; i<5; i=i+1){
		ask q1;
		for(j=0; j<3; j=j+1){
			ask q2;
		}
	}
}

Here q2 is at a deeper nesting level than q1, and both are array questions
when we reach q1[1] (2nd iteration) q2[0], q2[1], q2[2] are already taken
and at q1[1] we need to save these questions into the scope.

The mechanism to facilitate this is now put in place by GetQuestionsInBlock
along with helper operators which have been made friends of the AbstractQuestion class
listed below for reference:

	friend bool IsInTheSameScopeAndLevel(AbstractQuestion *q1, AbstractQuestion * q2);
	friend bool NotInTheSameBlock(AbstractQuestion *q1, AbstractQuestion * q2);
	friend bool IsAtAHigherNestLevelInTheSameBlock(AbstractQuestion *q1, AbstractQuestion * q2);
	friend bool IsAtADeeperNestLevelInTheSameBlock(AbstractQuestion *q1, AbstractQuestion * q2);

	they are written with the help of AbstractQuestion class member function 
	bool QuestionIsInMyBlock(AbstractQuestion *q);

The next set of changes will use the Framework above to save array questions 
depending on how they occur in various blocks.