small modification needed to lexer for stubs/inp_combined_bcpp.C
inp_combined_bcpp.C gives us some web output right now
It will be shaped into something that works on the web
and then the compiler will generate the same code.

scan_data.l has been modified because we needed a support
function.

At this point I would like to point out as a note that
scan_data.l will need to be made a pure - parser
i.e. multiply re-entrant so that it works on the web
currently we have vector<int32_t> data which is a global
variable that holds the parsed results.

This works in our linear flow in ncurses but will not work in the
web server