new named range syntax with support for groups
we allow nested groups.

random_test.input is not valid but for historic purpose
random_test2.input demonstrates the new syntax which is correctly
parsed. The new syntax supports nested groups, and printing shows
that the group gets built correctly.

My next focus will be on the randomizer - with support for
randomization withing groups and keeping groups together