improved qax generation
qax generation has a long way to go. This is an initial attempt
that will only work in special cases. The entire qax generation
engine has to be re-written using a stack to store if conditions
that come along the way. This just marks an initial milestone on the
journey