read question text as message from xml file
This is preparation for internationalization.
This was just proof of concept, but not what we want.

The struct TextExpression in question.h
will be modified to contain a pointer to the named attribute list
At runtime the named attribute will be loaded rather than
dumping it into a message in the messages.xml (internationalization file)