reading of windows registry
This program is key to modifying the compiler for windows.
Considering how complicated the install is, i would like to create an installer
that installs without any user intervention if possible.

This means that I need to automatically set the path to the g++ compiler,
the QSCRIPT_HOME directory etc. Setting these values in the registry is a good
starting point. Since I have never worked with the windows registry before,
looks like the day has finally arrived :