find language translation files for jobs
This will become a function in the generated code. It will
populate a drop down widget at the start of the survey asking the
user to choose her language. It will then set the locale for the
survey.

I guess this means that I should also store the locale in a question
in the survey.