fixed bug in setadd
if q1 is single coded and we do a setadd,
we should first clear out q1 and then add the new value

There is also a special case for hidden questions.

the clear statement only clears out the isAnswered_ flag in questions,
but does not delete their data.
if a hidden question is set using previous questions, the input data
which is not cleared can cause interference and wrong values.
This commit fixes those problems

the following case in setadd_bug.qscript demonstrates this bug