READ_EQ_1 mode - nicemap newcard decision: IMPORTANT - HowToFixNewCardBug
in READ_EQ_1 mode it doesnt make sense to just dump new
iterations data into a new card - it would be a waste of cards

instead add about 100 to the current location and if we are at a
card boundary then do a new card

There is a bug in the newcard generation

for (...) {
q1 ...
q2 ...
q3 ...
}

for the 1st iteration in the above case q1.0, q2.0, q3.0 all
go into new cards, it should be that only q1.0 goes into a new
card.

The way to fix this bug is keep a pointer to the previous question
in the current question for which we are generating the map,
and check if it belongs in the same for block as us and is
also iteration 0. If yes then dont do the newcard for q2.0 etc.