headed towards randomization code generation
After parsing the groups they are currently stored as vectors
which are chained together.

This step puts all group items into an array. This means that
an ordinary stub or a group pointer - all go into a single array
i.e. each element of a group occupies a unique position in the array
This paves the way for invoking random shuffle on the array and each
sub group of the stub list recursively.

this is the current programming map for this part of the project