randomization display is partially done
still need to get rid of all the FIXMEs in the code,
but we can now display the randomized text on the screen.

Porting this change to the web should not be a big deal,
as the core algorithms remain the same.

There is still a major part left, which is the
changes to writing the data files to the disk.