merged read random data grammar with original read data file grammar
This is an unclean commit. The qscript_data.ypp file was a
re-entrant grammar, it is not re-entrant any more and has to
be cleaned up. However, as of now, we can read the randomized
order stored on disk.

Significant changes are to be made, for example adding another
parameter to the bison/yacc parser to store the randomized order
data read from the disk. All global variables will also have to
be made as parameters.