work on array question recode
The test_fix_and_recode_nxd files are the iterations as we
progress towards the final version of the recode.

A lot more progress has been done and we are able to capture all
the array questions in the map. We now need to generate the definitions
and recode for these array questions