fixed code generation for ARRAY questions: bug still exists
This fixes the recode-edit qax code generation for single level
array questions. For 2 level loops or more , this problem is not fixed.

I have noticed that the number of axes generated in this case exceeds
quantum's limits and we will need to run the tables generated in parts
i.e if the number of axes > 500