1st attempt to make qscript_data.ypp re-entrant
qscript_data.ypp is the parser that reads entered data from the data
file. Now that we want randomization for web and cati, we need to
store the randomization order to disk and then be able to read the
same randomization for the respondent who returns back to a survey.

For the data reader to work in a web environment, the parser should
be fully re-entrant, i.e. no local variables.

Hopefully, I got it right in the 1st go.