create a default radio button, selected not displayed
This solves the problem of default button being pre-selected
in the radio button group. If we dont choose one, gtk chooses
one for us. By not displaying the rb, it appears as if nothing
is selected.

Need to setup a handler that checks if the selected button is
not the one with code 0