partial bug fix for incorrect if/else code generation of array questions
was generating incorrect code for the if/else for clearing out the array
question. This problem is only handled in specific cases, for other
cases the code generation is still probably wrong.

I have tested the cases and added a new case , which I will add in the
next commit. The code generated seems correct for these cases.

However, the 1st next step, is look at the nested scope stack stored in
the compound statement and detect interleaved if / else statements with
for loops. This is primarily where the error occurs and I need to
first detect this and inform users that this case has not been handled
correctly by the compiler and exit with an error