bug fix : code generation with interleaved if and for statements
this fix is not proper - for_with_if7.qscript is the test case.
Right now I am not keeping a track of the nested for level in
the compound statement. Hence, the height value can be wrong
when there are 2 if statements conditions one on top of the other

like this

for loop ...
	if (code in q1)
		if (code2 in q2)
q2_2               "q2_2"

	this will be corrected in the next fix, once i setup the
	test case correctly