incremental improvement to struct code generation
the current generated code will not compile.
This is a progression towards fast compilation times
1. reuse structs which have a common include file
   the include file is a hint passed by qscript
   when generating the axes

2. transfer all struct defns into .h without any function
   bodies and put func defns only into .cpp file

My current attack is on the generate_axes_code function
normalizing all the bullshit that I did there because I was
in a hurry to get things to work. The clean up is slow,
tedious and painful, but we march ahead