fixed web gui for multiple sessions
the run and compile functions need to be moved
to generate_code.cpp file and also need to be cleaned up.
I have put an early return in one of the functions above so
that I dont have to bother with compilation on windows,
but this has to be fixed in the long run.