there is some doubt about the ordering for columns and rows
This function has to be cleared up by matching the output
got by the simple version and the parallel version.

Right now, I am not clear as what is stored as rows - banner data
or side data.

Need to review the print function in the generated code for
tables to understand this.