dead end for trying to make gtk and ncurses look alike
This is a note I have put down in a note book and logging it here as
well.

I have been trying to make the ncurses generated code and the
gtk (and Wt) look alike. This wont work however, because

1. ncurses is not evented
2. gtk (and Wt) are

In a non-evented system - the program flow is sequential

In an evented system the flow is triggered by a call back. There is
already an event loop running, just that somehow you can forget it's
there.

I would have liked the gtk code to look like the ncurses code, which is
how it would probably look in Erlang (without the destructive updates of
course), but as I have realised not possible.

What i can do to make them match the best is, make both the functions as
similar as possible and make sure I handle all the cases.

My new task is to separate out struct TheQuestionnaire into 2
components, the Abstract part encapuslates all the functions that can be
done wihout the code being generated dynamically.

I have had to learn this evented/sequential lesson 2 times
unfortunately. hopefully by making a log message like this I will not
repeat it again.

My second mistake has been not focusing on the runtime early enough and
being ok with ugly looking code that was not normalised for a long time.
I should try and make the runtime code as clean and beautiful as
possible