identified why radio is not updated
the getline call in GetInputFromUser blocks the wx events.
Hence I need to first get rid of the getline call
and take all the data from the wx UI.

this means comparing the gtk2 runtime and the wx runtime and making them
as similar as possible. that is the next task