initial ui frame for invoking the compiler
The objective is to make this platform independent
After analysing the problems on windows - the default environment is
only 32kb. Writing an installer could clobber user variables. For
example my own installation of windows has:
 haskell, python, ruby (atleast 2 versions), erlang, various lisps and
 god knows what else. All these have been added to my path variable.

 Hence the plan is to distribute  this as a zip file and it will execute
 the compiler by understanding it's current working dir and invoke the
 compiler correctly.

It will also check the environment variable size and inform the user of
the problem it is facing - esp if this is more than 32kb.

Of course none of this functionality is implemented, but hope to have
it all working soon.