fixed memory bug on program exit but I dont know why it works
Need to investigate it further. When I invoke the create function in the
constructor it does not crash on exit. otherwise (comment it out) and it
will crash. There was some memory allocation that it was expecting, and
by not calling the function, I was preventing that from happening
resulting in the crash. Too tired to figure it out, but at least i can
sleep knowing that the bug is not there :-)