output modified for new runtime changes
We have a long way to go incorporating the changes in question.h
back into the mainline compiler will probably be a nightmare.

However separating out the runtime and compile time are crucial to
multi-platform running exes.

As stated in the README in this dir , we want to remove dependencies
like ncurses and Wt etc which came in from the old code base.

our objective is to make separate runtime files for each library, and
stdout being the simplest runtime possible.

The real fun will start when I go to AbstractQuestion (which is now
commented out with a #if 0) and replace the class variables with a
AbstractRuntimeQuestion * ptr . That is the nightmare I have to face :
coming up tomorrow :-)