works with emscripten .
Bug was due to early emscripten which does not handle cin properly
(see interactive_hello*.cpp series)

see emcc version below (emcc --version)
emcc (Emscripten GCC-like replacement) 1.3.2 (commit a8e4801c7d38033fff760ea26a4579aa324e303e)

The work around was changing cin to fgets with a fixed buffer.
Even cin.get did not work.
I commented out all logging to qscript_stdout, maybe that was not
necessary. Hence I have split out the files into src-emscripten so that
this can be tested independently.

There was also 1 bug in my code in a function that was supposed to
return a value, but it was not -
Also vegetable.cpp has been modified (parse_options) has been commented
out. It could be that parse_options is also not returning a value at the
end and may need to be re-looked at.

The flex scanner embedded in this worked fine - and surprised me.
My next steps would be invoking some pure javascript functions in an
embedded file - like jquery.js or dojo.js and see if that works
This is possible in emscripten as per the docs
And of course we need to transmit the survey data using a POST.

Once this is done, replace jqueryui with phonegap/sencha touch and we
are good to go!