previous version causing errors in FF. ran in Chrome with errors
When replacing the stubs in the form (radio boxes at present) with
new stubs , if the same list was being created with the same name , we
were having an error with the dojo parser, because it was trying to
register the same widget name.

I have passed in a counter to the js function as a hack to create a
unique name. However this is not a good option, the best one would be to
create and object cache - when we do a remove child, remove the entire
list of dom nodes and dump it into the object cache hashed by the name
of the stub list. Then when the same list comes up again, first look up
in the cache and then restore it. The only problem that will come up is
related to masking the for stubs that have setadd/setdel working on
them. That is why I have not bothered with the caching method for now.
Just get it to work and we will optimize for speed and less consing
later