removed further dependencies from runtime. active_var_info.h may be required
not sure of active_var_info is required in the runtime when compiling
without -o .

The data file creation support has been removed. We can create a
separate runtime for that. This simplifies the survey programming script
a lot more and the compile time will be a lot faster with emscripten