display single/multi-coded stubs on screen
We are not yet handling the response, that will come up next.

This whole exercise is about arranging the prev wt code in the new
format, matching the layout of the wx and gtk and stdout runtimes.

Once the entire wt functionality is implemented, we need to add any new
functionality (for example loading and saving data) back into gtk,
stdout and wx.

Then we set our sights back on emscripten.