call cpp function from qnre
This is one up from calling a cpp function from a simple example.
I got stuck because I forgot the leading "_" in -s EXPORT_FUNCTION.

Other things were, i had to export _main. This was a guess, because the
program was not running, and exporting _main worked. Not sure if this is
right.

Reached a stage when from inside the called_back_from_dom function I
want to invoke the question/questionnaire. But I cannot, since this is
outside a class.

Hence tomorrow, I plan on adding a static function to
AbstractQuestionnaire which will give me the whole questionnaire.

Since this is html running in a browser and for single use, I dont think
the global variable will cause a problem (I hope not)

The next experiment is hard code a value and pass that to the equivalent
next function (taking our cue from question_wx2_runtime/question_wt_runtime)
and implmenting it here. The objective is to see that the next question
loads in the browser ( because I dont know what happens to the cpp
code/unloading of the object from memory etc).

There are notes in the emscripten_api_browser_infloop.cpp that the
desctructor should not have been called on a class. I am relying on this
behaviour as "main" has exited after setting up the 1st question.