check box now displayed. data transferred back to callback
We are inching forward. Need to add a text box to capture RangeQuestion
data.

Need to pass RangeQuestion upper/lower bounds or valid inputs to the
frontend so that we can use some function there to apply validations.
Not that these checks wont be done at the back end.

Also noticed that once emscripten exits main, cout does not work, but
printf works. Is this some emscripten bug, or maybe Im using it wrong?