dojo works fine now.
It took me a while to understand how to put it together.
I was invoking the dojo parser incorrectly, the documentation had
exactly what I wanted: how to call the parser on a specific subtree of
DOM nodes.
There is a warning showing about dijit/form/Form being auto required.
Need to require this explicitly.

Tomorrow's attack will be on Dojo Mobile - with all the new learnings in
place. Hopefully it should work well