fixed many registry issues
The work in handling registry keys was left abandoned for a long time.
Decided today to clean up the WIN installer, which requires that the
retrieval of values from the WIN registry should be working properly.

There is more testing to be done. And Maybe I should add some checksum
features to be sure that corrupted exes cannot be distributed etc.

Currently we are getting positives on the sanity checks - seems to be fine.