improved - but not there yet
we are currently displaying the randomized output.
the test_randomization_1.cpp file is my manual changes which
will be merged back into the code generator.

Currently there is a mistake in VectorizePrint - I am storing the stub
codes into the array, instead i should be storing the index of the stub
if it were flattened out in the array

The next few code commits will correct this. First, we need to create a
nice test case which will help us get there faster