PARTLY FIXED BUG in questionIndexNo_ - FIX LATER
This is a bug I introduced while generating code for questionIndexNo_
earlier. The problem was, the survey kept going back to questions
previously answered - but only if they were looped questions.

I tracked this down to questionIndexNo_, array questions are created
later, and had an index no which was after the last quesion in the
survey. Due to the need for handling the PREVIOUS question,
we had introduced questionIndexNo_, but array questions were being
assigned positions after the last question in the survey and causing
the logic condition of PREVIOUS question to be triggered.

This is a fix which will not work for loops which have 2 or more levels
of nesting. Hence I have printed this out in the output. Also there is
an implicit, assumption that a looped question will have less than 1000
iterations. This is a bad fix and needs to be re-visited, when I am
fresh and have more brain power at my disposal