sample android phonegap app
- replace index.html  in assets/www with the html file from the compiler
  output. For example

simple_compiler> cd inputs
inputs> ../bin/qscript_wq2    -o -m --emscripten -f vegetable-old.qscript

 cp vegetable-old.html ../../qscript_android_phonegap_app/assets/www/index.html

Currently the committed shell file shell-phonegap...html
is trying to open the data file for the serial no entered in a hardcoded
manner. I am trying to fix this behaviour but unable to solve the bug,
some issue in the filesystem api maybe? as the fail function does not
execute when it is supposed to. Need to figure out a way to solve this
problem.