multiple question running for simple case - still needs to be fixed
So have currently hacked - is the word - and managed to get the
generated code to compile. It will only work for questions which are not
in a for loop - as array question code generation has not been touched.

I am currently just taking the 1st question in a page and showing it.
This obviously needs to be cleaned up.

There are many assumptions in the code about there being just one
question on display. All these points will now have to be looked at.

Examples:
	One set of radio box holder, check box holder.

	Instead we will now need a UI template, one for each question
	returned by the eval2 function.

	Then all the forms (ui templates) will be filled and dumped into
one grand container, which is what we display on screen.

	Also need to clean up the function that handles the response
from the system. If there are multiple questions then obviously it
should not return success unless there is a valid answer for all
questions on page.

	Also need to report error messages on screen, have not handled
that as yet.