partial commit: displays stubs on screen for multiple questions
we are slowly getting there. Next step, modify the handler which
sends the answers to the backend to gather the data from all the forms
which contain stubs and accumulate the answers.

Currently the ui does not have separate forms for each question/stub
list. We now need to look into this.